requirements_qns.txt
absl-py==1.3.0
appdirs==1.4.4
attrs==22.2.0
autograd==1.5
autoray==0.5.3
backcall==0.2.0
brotlipy==0.7.0
cachetools==5.2.0
carla==0.9.13
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.3
cloudpickle==2.2.0
colorama==0.4.6
ConfigArgParse==1.5.3
cryptography==38.0.1
cycler==0.11.0
dash==2.7.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.4
decorator==5.1.1
dpcpp-cpp-rt==2022.2.1
entrypoints==0.4
fastjsonschema==2.16.2
Flask==2.2.2
flit_core==3.6.0
fonttools==4.38.0
future==0.18.2
google-auth==2.15.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
gym==0.26.2
gym-notices==0.0.8
idna==3.4
importlib-metadata==5.2.0
importlib-resources==5.10.1
intel-cmplr-lib-rt==2022.2.1
intel-cmplr-lic-rt==2022.2.1
intel-opencl-rt==2022.2.1
intel-openmp==2022.2.1
ipykernel==6.16.2
ipython==7.34.0
ipywidgets==8.0.4
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==7.4.8
jupyter_core==4.12.0
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mkl==2022.2.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
nbformat==5.5.0
nest-asyncio==1.5.6
networkx==3.0
ninja==1.11.1
numpy==1.22.3
oauthlib==3.2.2
open3d==0.16.0
opencv-python==4.6.0.66
packaging==22.0
parso==0.8.3
PennyLane==0.28.0
PennyLane-Lightning==0.28.1
pickleshare==0.7.5
Pillow==9.3.0
pip==22.3.1
pkgutil_resolve_name==1.3.10
plotly==5.11.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pygame==2.1.2
Pygments==2.13.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.19.2
PySocks==1.7.1
python-dateutil==2.8.2
pywin32==305
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
retworkx==0.12.1
rsa==4.9
rustworkx==0.12.1
scikit-learn==1.0.2
scipy==1.7.3
semantic-version==2.10.0
setuptools==65.5.0
six==1.16.0
tbb==2021.7.1
tenacity==8.1.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
toml==0.10.2
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.2
traitlets==5.8.0
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
Werkzeug==2.2.2
wheel==0.37.1
widgetsnbextension==4.0.5
win-inet-pton==1.1.0
wincertstore==0.2
zipp==3.11.0

# python_v = 3.8